Refactoring of the debug framework. Move compiler thread implementation to its own file.

+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.graal.debug/src/com/oracle/graal/debug/internal/MetricImpl.java
@@ -34,7 +34,7 @@ public final class MetricImpl extends DebugValue implements DebugMetric {
         add(1);
     }
 
-    public void add(long value) {
+    public void add(int value) {
         if (Debug.isMeterEnabled()) {
             super.addToCurrentValue(value);
         }
